# PaymentGateway/PaymentGateway.API/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Payment Gateway",
    "description": "Gateway to facilitate processing payments and to retrieve payment details.\n",
    "contact": {
      "name": "Kamal Mahmassani",
      "email": "kmahmassani@gmail.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
    "/payments": {
      "post": {

        "tags": [ "Payments" ],

        "summary": "Request a payment",
        "description": "To verify the success of the payment, check the `approved` field in the response.\n",
        "parameters": [
          {

            "name": "Authorization",
            "in": "header",
            "description": "Your valid merchant account secret key",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Content-Type",


            "in": "header",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string",

              "enum": [ "application/json" ]


            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Payment processed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Invalid data was sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "502": {
            "description": "Bad gateway"
          }
        }
      }
    },
    "/payments/{id}": {
      "get": {

        "tags": [ "Payments" ],

        "summary": "Get payment details",
        "description": "Returns the details of the payment with the specified identifier string.\n\nIf the payment method requires a redirection to a third party (e.g., 3D Secure),\nthe redirect URL back to your site will include a `cko-session-id` query parameter\ncontaining a payment session ID that can be used to obtain the details of the payment, for example:\n\nhttp://example.com/success?cko-session-id=sid_ubfj2q76miwundwlk72vxt2i7q.\n",
        "parameters": [
          {

            "name": "Authorization",
            "in": "header",
            "description": "Your valid merchant account secret key",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          },
          {



            "name": "id",
            "in": "path",
            "description": "The payment or payment session identifier",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "pattern": "^(pay|sid)_(\\w{26})$",
              "type": "string"

            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Payment not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "PaymentRequest": {
        "properties": {
          "source": {
            "$ref": "#/components/schemas/PaymentRequestSource"
          },
          "amount": {
            "minimum": 0,
            "type": "integer",
            "description": "The payment amount.\nThe exact format <a href=\"https://docs.checkout.com/docs/calculating-the-value\" target=\"blank\">depends on the currency</a>.\n",
            "example": 6540
          },
          "currency": {
            "maxLength": 3,
            "minLength": 3,
            "type": "string",
            "description": "The three-letter <a href=\"https://docs.checkout.com/docs/currency-codes\" target=\"blank\">ISO currency code</a>\n",
            "example": "USD"
          },
          "reference": {
            "type": "string",
            "description": "A reference you can later use to identify this payment, such as an order number.",
            "example": "ORD-5023-4E89"
          }
        }
      },
      "PaymentRequestSource": {
        "type": "object",
        "description": "A card payment source",
        "allOf": [
          {
            "required": [ "expiry_month", "expiry_year", "number", "type" ],






            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "description": "The type of payment source. Set this to `card`.",
                "example": "card"
              },
              "number": {
                "type": "string",
                "description": "The card number (without separators)",
                "example": "4543474002249996"
              },
              "expiry_month": {
                "minimum": 1,
                "maxLength": 2,
                "minLength": 1,
                "type": "integer",
                "description": "The expiry month of the card",
                "example": 6
              },
              "expiry_year": {
                "maxLength": 4,
                "minLength": 4,
                "type": "integer",
                "description": "The expiry year of the card",
                "example": 2025
              },
              "name": {
                "type": "string",
                "description": "The name of the cardholder",
                "example": "Bruce Wayne"
              },
              "cvv": {
                "maxLength": 4,
                "minLength": 3,
                "type": "string",
                "description": "The card verification value/code. 3 digits, except for Amex (4 digits)",
                "example": "956"

              }
            }
          }
        ]

      },
      "PaymentResponse": {
        "required": [ "_links", "amount", "approved", "currency", "id", "processed_on", "status" ],








        "type": "object",
        "properties": {
          "id": {
            "description": "The payment's unique identifier",
            "allOf": [
              {

                "$ref": "#/components/schemas/PaymentId"
              }
            ]


          },
          "amount": {
            "type": "integer",
            "description": "The payment amount",
            "example": 6540
          },
          "currency": {
            "maxLength": 3,
            "minLength": 3,
            "type": "string",
            "description": "The three-letter <a href=\"https://docs.checkout.com/docs/currency-codes\" target=\"blank\">ISO currency code</a> of the payment",
            "example": "USD"
          },
          "approved": {
            "type": "boolean",
            "description": "Whether or not the authorization or capture was successful",
            "example": true
          },
          "status": {




            "type": "string",
            "description": "The status of the payment",
            "example": "Authorized",
            "enum": [ "Authorized", "Declined" ]
          },
          "auth_code": {
            "type": "string",
            "description": "The acquirer authorization code if the payment was authorized",
            "example": "643381"
          },
          "source": {
            "type": "object",
            "description": "The source of the payment",
            "allOf": [
              {

                "$ref": "#/components/schemas/PaymentResponseSource"
              }
            ]


          },
          "processed_on": {





            "description": "The date/time the payment was processed",
            "allOf": [
              {
                "$ref": "#/components/schemas/ServerTimestamp"
              }
            ]
          },
          "reference": {
            "type": "string",
            "description": "Your reference for the payment",
            "example": "ORD-5023-4E89"
          },
          "_links": {










            "$ref": "#/components/schemas/PaymentResponse__links"











          }
        },
        "description": "Payment Response",
        "example": {
          "reference": "ORD-5023-4E89",
          "amount": 6540,
          "approved": true,
          "_links": "{\"self\":{\"href\":\"https://api.checkout.com/payments/pay_y3oqhf46pyzuxjbcn2giaqnb44\"}}",
          "currency": "USD",
          "id": "",
          "source": "",
          "processed_on": "",
          "status": "Authorized",
          "auth_code": "643381"
        }
      },
      "PaymentResponseSource": {
        "type": "object",
        "description": "A card payment source",
        "allOf": [
          {
            "required": [ "bin", "expiry_month", "expiry_year", "fingerprint", "last4" ],







            "type": "object",
            "properties": {
              "expiry_month": {
                "minimum": 1,
                "maxLength": 2,
                "minLength": 1,
                "type": "integer",
                "description": "The expiry month",
                "example": 6
              },
              "expiry_year": {
                "maxLength": 4,
                "minLength": 4,
                "type": "integer",
                "description": "The expiry year",
                "example": 2025
              },
              "name": {
                "type": "string",
                "description": "The cardholder's name",
                "example": "Bruce Wayne"
              },
              "scheme": {
                "type": "string",
                "description": "The card scheme",
                "example": "VISA"
              },
              "last4": {
                "type": "string",
                "description": "The last four digits of the card number",
                "example": "9996"

              }
            }
          }
        ]

      },
      "Link": {
        "required": [ "href" ],


        "type": "object",
        "properties": {
          "href": {
            "type": "string",
            "description": "The link URL"
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "request_id": {
            "type": "string",
            "example": "0HL80RJLS76I7"
          },
          "error_type": {
            "type": "string",
            "example": "request_invalid"
          },
          "error_codes": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "payment_source_required"
            }
          }
        }
      },
      "PaymentId": {
        "maxLength": 30,
        "minLength": 30,
        "pattern": "^(pay)_(\\w{26})$",
        "type": "string",
        "description": "The payment identifier",
        "example": "pay_mbabizu24mvu3mela5njyhpit4"
      },
      "ServerTimestamp": {



        "type": "string",
        "description": "Read-only UTC timestamp, automatically assigned by us",
        "format": "date-time",
        "readOnly": true
      },
      "PaymentResponse__links": {
        "required": [ "self" ],
        "type": "object",
        "properties": {
          "self": {
            "type": "object",
            "description": "The URI of the payment",
            "allOf": [
              {
                "$ref": "#/components/schemas/Link"
              }
            ]
          }
        },
        "description": "The links related to the payment",
        "example": "{\"self\":{\"href\":\"https://api.checkout.com/payments/pay_y3oqhf46pyzuxjbcn2giaqnb44\"}}"
      }
    }
  }
}